tests/requirements-test.txt
pytest==8.3.5
pytest-timeout==2.3.1
redis==5.2.1
psycopg2-binary==2.9.10
neo4j==5.28.1
requests==2.32.3
python-dotenv==1.0.1
